atc/component: structure: Scheduler -> Runner
