+++ atc/component/component.go
@@ -1,0 +1,12 @@
+package component
+
+type Component interface {
+	Name() string
+	Paused() bool
+
+	IntervalElapsed() bool
+
+	Reload() (bool, error)
+	UpdateLastRan() error
+}
+
